requirements.txt
streamlit
pandas
matplotlib
pillow
numpy
chardet
nltk